# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: $MAIL — Solana wallet mailbox API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blocked:
    get:
      operationId: listBlockedWallets
      tags: [safety]
      summary: List wallets blocked by the signed-in wallet
      responses:
        "200":
          description: Blocked wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedWallet"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    post:
      operationId: blockWallet
      tags: [safety]
      summary: Block a sender wallet
      description: Existing mail from the sender moves to Junk. Future mail is refused.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockInput"
      responses:
        "201":
          description: Wallet blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedWallet"
        "400":
          description: Invalid wallet address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /blocked/{wallet}:
    delete:
      operationId: unblockWallet
      tags: [safety]
      summary: Unblock a sender wallet
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /reports:
    post:
      operationId: reportSender
      tags: [safety]
      summary: Report a message/sender
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    BlockInput:
      type: object
      properties:
        wallet:
          type: string
      required:
        - wallet
    BlockedWallet:
      type: object
      properties:
        wallet:
          type: string
        createdAt:
          type: string
      required:
        - wallet
        - createdAt
    ReportInput:
      type: object
      properties:
        messageId:
          type: string
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - messageId
        - reason
    Report:
      type: object
      properties:
        id:
          type: string
        messageId:
          type: string
        reason:
          type: string
        createdAt:
          type: string
      required:
        - id
        - messageId
        - reason
        - createdAt
...